requirements.txt
# EmberGuide POC Requirements

# Web Frameworks
fastapi==0.104.1
uvicorn[standard]==0.24.0
streamlit==1.28.2
streamlit-folium==0.15.1
pydantic==2.5.0

# Geospatial Processing
rasterio==1.3.9
# GDAL==3.8.0  # Included with rasterio/geopandas on Windows
pyproj==3.6.1
shapely==2.0.2
geopandas==0.14.1
folium==0.15.0

# Weather Data
xarray==2023.11.0
netCDF4==1.6.5
cdsapi==0.6.1

# Scientific Computing
numpy>=1.26.0
scipy==1.11.4
pandas==2.1.3
scikit-learn==1.3.2

# Clustering & ML
scikit-learn==1.3.2
joblib==1.3.2

# Configuration & Utilities
pyyaml==6.0.1
python-dotenv==1.0.0
requests==2.31.0
click==8.1.7

# Logging & Monitoring
structlog==23.2.0

# Testing (optional for POC)
pytest==7.4.3
pytest-cov==4.1.0

